requirements.txt
requests==2.12.3
pyOpenSSL==16.2.0
